% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_info.R
\name{default_models}
\alias{default_models}
\title{Model help}
\usage{
default_models()
}
\value{
data frame of model descriptions.
}
\description{
This function prints a list of all built in models, along with their parameter arguments and a short description.
}
\examples{
default_models()
}
