% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_ics.R
\name{get_ics}
\alias{get_ics}
\title{Get default initial conditions}
\usage{
get_ics(model)
}
\arguments{
\item{model}{name of the model to be solved. Examples include: SIR and SIR vaccination.}
}
\value{
named vector of default initial conditions.
}
\description{
This function returns the default parameter vectors for a particular built-in model.
}
