% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_names.R
\name{get_name}
\alias{get_name}
\title{Get model display names}
\usage{
get_name(model)
}
\arguments{
\item{model}{character specifying the name of the built-in model.}
}
\value{
character of the corresponding display name.
}
\description{
This function gets the display name for any built-in model.
}
\examples{
get_name(model = "SIR")
}
