% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params.R
\name{get_params}
\alias{get_params}
\title{Get default parameters}
\usage{
get_params(model)
}
\arguments{
\item{model}{name of the model to be solved. Examples include: SIR and SIR vaccination.}
}
\value{
list of default parameter vectors.
}
\description{
This function returns the default parameter vectors for a particular built-in model.
}
