% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{SIR}
\alias{SIR}
\title{SIR model}
\usage{
SIR(t, y, parms)
}
\arguments{
\item{t}{numeric vector of time points.}

\item{y}{numeric vector of variables.}

\item{parms}{named vector of model parameters.}
}
\value{
equation list
}
\description{
These equations describe the classic SIR model with no births or deaths.
}
