% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{SIRbirths}
\alias{SIRbirths}
\title{SIR model with demography}
\usage{
SIRbirths(t, y, parms)
}
\arguments{
\item{t}{numeric vector of time points.}

\item{y}{numeric vector of variables.}

\item{parms}{named vector of model parameters.}
}
\value{
equation list
}
\description{
These equations describe the classic SIR model with equal births and deaths.
}
