% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{SIRS}
\alias{SIRS}
\title{SIRS model
These equations describe the classic SIRS model without births or deaths.}
\usage{
SIRS(t, y, parms)
}
\arguments{
\item{t}{numeric vector of time points.}

\item{y}{numeric vector of variables.}

\item{parms}{named vector of model parameters.}
}
\value{
equation list
}
\description{
SIRS model
These equations describe the classic SIRS model without births or deaths.
}
