% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{SIRSbirths}
\alias{SIRSbirths}
\title{SIRS model with demography}
\usage{
SIRSbirths(t, y, parms)
}
\arguments{
\item{t}{numeric vector of time points.}

\item{y}{numeric vector of variables.}

\item{parms}{named vector of model parameters.}
}
\value{
equation list
}
\description{
These equations describe the classic SIRS model with equal birth and death rates.
}
