% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{SIRSvaccination}
\alias{SIRSvaccination}
\title{SIRS model with vaccination at birth
These equations describe the classic SIRS model with equal birth and death rates and vaccination at birth.}
\usage{
SIRSvaccination(t, y, parms)
}
\arguments{
\item{t}{numeric vector of time points.}

\item{y}{numeric vector of variables.}

\item{parms}{named vector of model parameters.}
}
\value{
equation list
}
\description{
SIRS model with vaccination at birth
These equations describe the classic SIRS model with equal birth and death rates and vaccination at birth.
}
