% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{SIRvaccination}
\alias{SIRvaccination}
\title{SIR model with vaccination at birth
These equations describe the classic SIR model with births and deaths, constant population size, and (optional) vaccination at birth.}
\usage{
SIRvaccination(t, y, parms)
}
\arguments{
\item{t}{numeric vector of time points}

\item{y}{numeric vector of variables}

\item{parms}{named vector of model parameters.}
}
\value{
equation list
}
\description{
SIR model with vaccination at birth
These equations describe the classic SIR model with births and deaths, constant population size, and (optional) vaccination at birth.
}
