% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{SIS}
\alias{SIS}
\title{SIS model
These equations describe the classic SIS model with no births or deaths.}
\usage{
SIS(t, y, parms)
}
\arguments{
\item{t}{numeric vector of time points}

\item{y}{numeric vector of variables}

\item{parms}{named vector of model parameters.}
}
\value{
equation list
}
\description{
SIS model
These equations describe the classic SIS model with no births or deaths.
}
