% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solving.R
\name{solve_eqns}
\alias{solve_eqns}
\title{Solve equations}
\usage{
solve_eqns(eqns, ics, times, parms)
}
\arguments{
\item{eqns}{name of the model to be solved. Examples include: SIR and SIRvaccination.}

\item{ics}{named numeric vector specifying the initial conditions i.e. the initial values of all model variables.}

\item{times}{numerical vector indicating the time points at which the equation should be solved.}

\item{parms}{named numeric vector of parameter values.}
}
\value{
data frame of model solutions in long format.
}
\description{
This function solves an ODE model using 'deSolve' and returns the output as a data frame.
}
