% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_model}
\alias{plot_model}
\title{Plot model output.
This function plots the output of a fitted model data frame.}
\usage{
plot_model(
  output,
  linesize,
  textsize,
  xlabel,
  ylabel,
  legend_title,
  levels,
  values,
  ...
)
}
\arguments{
\item{output}{data frame output from solve_eqns().}

\item{linesize}{numeric value for line width in ggplot.}

\item{textsize}{numeric value for textsize in ggplot.}

\item{xlabel}{character string for x axis label.}

\item{ylabel}{character string for y axis label.}

\item{legend_title}{character string for legend title.}

\item{levels}{character vector of the variable names in the order they should be plotted. Default is to obtain the order from the initial conditions vector 'ics'.}

\item{values}{vector specifying manual color scale. Length must equal the number of model variables.}

\item{...}{extra argument to be passed through to ggplot scale_colour_manual: use 'labels' to change the legend names}
}
\value{
ggplot object
}
\description{
Plot model output.
This function plots the output of a fitted model data frame.
}
