% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{SIR}
\alias{SIR}
\title{SIR model without no births, deaths, or vaccination.
These equations describe the classic SIR model with constant population size, and no births or deaths.}
\usage{
SIR(t, y, parms)
}
\arguments{
\item{t}{numeric vector of time points.}

\item{y}{numeric vector of variables.}

\item{parms}{named vector of model parameters.}
}
\value{
equation list
}
\description{
SIR model without no births, deaths, or vaccination.
These equations describe the classic SIR model with constant population size, and no births or deaths.
}
